requirements.txt
flask
psycopg2-binary
bcrypt
pytest
requests